requirements.txt.txt
streamlit
google-analytics-data
google-auth-oauthlib
google-generativeai
pandas